Some actual useful progress?
I was able to make a file manager-ish thing, that writes text files to a
folder called 'TextFiles", and it allows you to name the file your self
within the program, and add the text via command line. It can also
overwrite any files that already exists, but doesn't show a list of the
ones that do [yet]. Also made the main class let you choose to run the
math class or the file class from the first screen.

This is still all in cmd though, not moved on to decent UI's yet.